First version of TTS Class
This is the initial version of the TTS class (v02) to discuss on our Friday meeting. "TestingTTSClass" is for testing the functionality. At the moment the idea is to have dialogue be added to a queue within the TTS class to then be popped upon request of the DM.
TO DOS:
- Major refactoring including successfully closing Threads
- Add to the queue every turn of conversation instead of a set time